+++ src/app/modules/marketplace/components/module-train/storage-conf-form/storage-conf-form.component.spec.ts
@@ -7,6 +7,9 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { SharedModule } from '@app/shared/shared.module';
 import { MediaMatcher } from '@angular/cdk/layout';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { AppConfigService } from '@app/core/services/app-config/app-config.service';
+import { ActivatedRoute } from '@angular/router';
+import { of } from 'rxjs';
 
 const mockedMediaQueryList: MediaQueryList = {
     matches: true,
@@ -22,6 +25,8 @@ const mockedMediaQueryList: MediaQueryList = {
 const mockedMediaMatcher: any = {
     matchMedia: jest.fn().mockReturnValue(mockedMediaQueryList),
 };
+
+const mockedConfigService: any = {};
 
 describe('StorageConfFormComponent', () => {
     let component: StorageConfFormComponent;
@@ -47,6 +52,21 @@ describe('StorageConfFormComponent', () => {
                 FormBuilder,
                 { provide: FormGroupDirective, useValue: formGroupDirective },
                 { provide: MediaMatcher, useValue: mockedMediaMatcher },
+                { provide: AppConfigService, useValue: mockedConfigService },
+                {
+                    provide: ActivatedRoute,
+                    useValue: {
+                        params: of({
+                            id: 'test',
+                        }),
+                        snapshot: {
+                            paramMap: {
+                                get: () => 'test', // represents the id
+                            },
+                        },
+                        routeConfig: { path: 'test' },
+                    },
+                },
             ],
         }).compileComponents();