+++ src/app/modules/marketplace/components/module-train/storage-conf-form/storage-conf-form.component.ts
@@ -13,6 +13,8 @@ import {
     confObjectStringBoolean,
 } from '@app/shared/interfaces/module.interface';
 import { ZenodoSimpleDataset } from '@app/shared/interfaces/dataset.interface';
+import { ProfileService } from '@app/modules/profile/services/profile.service';
+import { StorageCredential } from '@app/shared/interfaces/profile.interface';
 
 const mockedConfObject: confObject = {
     name: '',
@@ -32,6 +34,7 @@ const mockedConfObjectStringBoolean: confObjectStringBoolean = {
 })
 export class StorageConfFormComponent implements OnInit {
     constructor(
+        private profileService: ProfileService,
         private ctrlContainer: FormGroupDirective,
         private fb: FormBuilder,
         private changeDetectorRef: ChangeDetectorRef,
@@ -45,6 +48,10 @@ export class StorageConfFormComponent implements OnInit {
     parentForm!: FormGroup;
 
     storageConfFormGroup = this.fb.group({
+        storageServiceDatasetSelect: new FormControl({
+            value: '',
+            disabled: true,
+        }),
         rcloneConfInput: [''],
         storageUrlInput: [''],
         rcloneVendorSelect: [''],
@@ -109,8 +116,13 @@ export class StorageConfFormComponent implements OnInit {
     mobileQuery: MediaQueryList;
 
     hidePassword = true;
+    protected credentialsLoading = true;
+
     rcloneVendorOptions: { value: string; viewValue: string }[] = [];
+    protected storageServiceOptions: { value: string; viewValue: string }[] =
+        [];
     datasets: { doi: string; force_pull: boolean }[] = [];
+    credentials: StorageCredential[] = [];
 
     ngOnInit(): void {
         this.parentForm = this.ctrlContainer.form;
@@ -123,6 +135,8 @@ export class StorageConfFormComponent implements OnInit {
             this.storageConfFormGroup.get('rcloneUserInput')?.disable();
             this.storageConfFormGroup.get('rclonePasswordInput')?.disable();
         }
+
+        this.getLinkedStorageServices();
     }
 
     addDataset(dataset: ZenodoSimpleDataset): void {
@@ -161,6 +175,65 @@ export class StorageConfFormComponent implements OnInit {
         const d = this.datasets.find((d) => d.doi === dataset.doi);
         if (d) {
             d.force_pull = dataset.force_pull;
+        }
+    }
+
+    getLinkedStorageServices() {
+        this.profileService.getExistingCredentials().subscribe({
+            next: (credentials) => {
+                this.credentials = Object.values(credentials);
+                if (this.credentials.length > 0) {
+                    this.credentials.forEach(
+                        (credential: StorageCredential) => {
+                            this.storageServiceOptions.push({
+                                value: credential.server,
+                                viewValue: credential.server.replace(
+                                    'https://',
+                                    ''
+                                ),
+                            });
+                        }
+                    );
+                    this.storageConfFormGroup
+                        .get('storageServiceDatasetSelect')
+                        ?.setValue(this.storageServiceOptions[0].value);
+                    this.storageConfFormGroup
+                        .get('storageServiceDatasetSelect')
+                        ?.enable();
+                    this.credentialsLoading = false;
+                    this.updateStorageConfiguration();
+                } else {
+                    this.credentialsLoading = false;
+                }
+            },
+            error: () => {
+                this.credentialsLoading = false;
+            },
+        });
+    }
+
+    updateStorageConfiguration() {
+        const storageServiceUrl = this.storageConfFormGroup.get(
+            'storageServiceDatasetSelect'
+        )?.value;
+        const storageServiceCredentials = this.credentials.find(
+            (c) => c.server === storageServiceUrl
+        );
+        if (storageServiceCredentials) {
+            this.storageConfFormGroup
+                .get('rcloneUserInput')
+                ?.setValue(storageServiceCredentials.loginName);
+            this.storageConfFormGroup
+                .get('rclonePasswordInput')
+                ?.setValue(storageServiceCredentials.appPassword);
+            this.storageConfFormGroup
+                .get('storageUrlInput')
+                ?.setValue(
+                    storageServiceCredentials.server + '/remote.php/webdav/'
+                );
+            this.storageConfFormGroup
+                .get('rcloneVendorSelect')
+                ?.setValue(storageServiceCredentials.vendor);
         }
     }
 }